fix param.ipython auto import
