+++ param/depends.py
@@ -1,5 +1,3 @@
-import weakref
-
 from collections import defaultdict
 from functools import wraps
 
@@ -7,25 +5,6 @@ from .parameterized import (
     Parameter, Parameterized, ParameterizedMetaclass, transform_reference,
 )
 from ._utils import accept_arguments, iscoroutinefunction
-
-_display_accessors = {}
-_reactive_display_objs = weakref.WeakSet()
-
-def register_display_accessor(name, accessor, force=False):
-    if name in _display_accessors and not force:
-        raise KeyError(
-            'Display accessor {name!r} already registered. Override it '
-            'by setting force=True or unregister the existing accessor first.')
-    _display_accessors[name] = accessor
-    for fn in _reactive_display_objs:
-        setattr(fn, name, accessor(fn))
-
-def unregister_display_accessor(name):
-    if name not in _display_accessors:
-        raise KeyError('No such display accessor: {name!r}')
-    del _display_accessors[name]
-    for fn in _reactive_display_objs:
-        delattr(fn, name)
 
 
 @accept_arguments
